requirements.txt
numpy~=1.23.0
requests~=2.28.0
Pillow~=9.1.1
python-dotenv~=1.0.0
Flask~=2.1.2
scipy~=1.9.1